feat(checkEmailAvailability): allow email address from current contact
Prior to this fix, when editing a contact, the async validator
`checkEmailAvailability` invalidated the form because it checks
if the given email address (the one from the contact we're editing)
has already been taken, which is always the case because the server
checks against all email addresses of all existing contacts.

This commit extends the validator to take an optional whitelisted
email address (from a contact), so it stays valid when editing a
contact.