fix(ContactsEditor): only open canDeactivate dialog when model changed
Prior to this commit, the navigation guard didn't care if there
was a change in the contact model or not. This commit sets a property
to check if the model has actually been changed, which causes the
dialog to *not* open when we navigate away and no changes have been
done.